+++ compiler/common/PackageManifestSchema.ts
@@ -1,0 +1,2 @@
+export const BPL_PACKAGE_SCHEMA_URI =
+  "https://raw.githubusercontent.com/pr0h0/bpl3/master/bpl-package.schema.json";